Rewrite xtrabackup to support faunctionality previously available only for innobackupex.
