+++ storage/innobase/xtrabackup/test/inc/ib_local.sh
@@ -4,7 +4,7 @@
 # including:
 #
 # Optionally the following variables may be set:
-#   xtrabackup_option:  additional options to be passed to xtrabackup.
+#   innobackupex_option:  additional options to be passed to innobackupex.
 #   data_decrypt_cmd: command used to decrypt data
 #   data_decompress_cmd: command used to decompress data
 ############################################################################
@@ -16,12 +16,12 @@ start_server --innodb_file_per_table
 load_dbase_schema sakila
 load_dbase_data sakila
 
-xtrabackup_options=${xtrabackup_options:-""}
-xtrabackup_options="${xtrabackup_options} --no-timestamp"
+innobackupex_options=${innobackupex_options:-""}
+innobackupex_options="${innobackupex_options} --no-timestamp"
 
 # Take backup
 backup_dir=${topdir}/backup
-xtrabackup $xtrabackup_options --backup --target-dir=$backup_dir
+innobackupex $innobackupex_options $backup_dir
 vlog "Backup created in directory $backup_dir"
 
 stop_server
@@ -42,12 +42,12 @@ cd - >/dev/null 2>&1
 vlog "###########"
 vlog "# PREPARE #"
 vlog "###########"
-xtrabackup --prepare --target-dir=$backup_dir
+innobackupex --apply-log $backup_dir
 mkdir -p $mysql_datadir
 vlog "###########"
 vlog "# RESTORE #"
 vlog "###########"
-xtrabackup --copy-back --target-dir=$backup_dir
+innobackupex --copy-back $backup_dir
 
 start_server
 # Check sakila
