+++ storage/innobase/xtrabackup/test/t/bug817132.sh
@@ -17,7 +17,7 @@ load_dbase_data sakila
 
 mkdir -p $topdir/backup
 innobackupex  $topdir/backup
-backup_dir=`grep "innobackupex: Backup created in directory" $OUTFILE | awk -F\' '{ print $2}'`
+backup_dir=`grep "Backup created in directory" $OUTFILE | awk -F\' '{ print $2}'`
 vlog "Backup created in directory $backup_dir"
 
 stop_server
