+++ storage/innobase/xtrabackup/test/t/ib_basic.sh
@@ -1,0 +1,5 @@
+############################################################################
+# Test basic local backup
+############################################################################
+
+. inc/ib_local.sh
