+++ storage/innobase/xtrabackup/test/t/xb_stream_compress.sh
@@ -1,0 +1,12 @@
+############################################################################
+# Test streaming + compression
+############################################################################
+
+require_qpress
+
+stream_format=xbstream
+stream_extract_cmd="xbstream -xv <"
+stream_uncompress_cmd="innobackupex --decompress ./"
+xtrabackup_options="--compress --compress-threads=4 --compress-chunk-size=8K"
+
+. inc/xb_stream_common.sh
