+++ storage/innobase/xtrabackup/test/t/xb_stream_parallel.sh
@@ -1,0 +1,9 @@
+############################################################################
+# Test parallel streaming feature of the 'xbstream' format
+############################################################################
+
+stream_format=xbstream
+stream_extract_cmd="xbstream -xv <"
+xtrabackup_options="--parallel=16"
+
+. inc/xb_stream_common.sh
