+++ storage/innobase/xtrabackup/test/t/xb_stream_parallel_encrypt.sh
@@ -1,0 +1,12 @@
+############################################################################
+# Test streaming + parallel + encryption
+############################################################################
+
+
+encrypt_algo="AES256"
+encrypt_key="percona_xtrabackup_is_awesome___"
+stream_format=xbstream
+stream_extract_cmd="(xbstream -xv ; innobackupex --decrypt=$encrypt_algo --encrypt-key=$encrypt_key ./) <"
+xtrabackup_options="--parallel=16 --encrypt=$encrypt_algo --encrypt-key=$encrypt_key --encrypt-threads=4 --encrypt-chunk-size=8K"
+
+. inc/xb_stream_common.sh
