+++ storage/innobase/xtrabackup/test/t/xb_stream_tar.sh
@@ -1,0 +1,9 @@
+############################################################################
+# Test streaming in the TAR format
+############################################################################
+
+stream_extract_cmd="$TAR -ixvf"
+
+stream_format=tar
+
+. inc/xb_stream_common.sh
