+++ storage/innobase/xtrabackup/test/t/xb_stream_xbstream.sh
@@ -1,0 +1,8 @@
+############################################################################
+# Test streaming in the XBSTREAM format
+############################################################################
+
+stream_format=xbstream
+stream_extract_cmd="xbstream -xv <"
+
+. inc/xb_stream_common.sh
